show stars
why? because I guess they became relevant. Kinda annoying that it'll be wildly out of date unless I make it auto-rebuild but yeah, people will click it once in a while and see the real number so it doesn't matter that much. I'll make sure to trigger a rebuild when I do something significant.